refactor(ui): decompose terminal container view
